Move TokenTest to shared
