+++ sconfig/jvm/src/test/scala/org/ekrich/config/impl/ConfParserTest.scala
@@ -12,6 +12,7 @@ import org.ekrich.config._
 import scala.jdk.CollectionConverters._
 import java.net.URL
 import java.util.Properties
+import FileUtils._
 
 class ConfParserTest extends TestUtils {
   def parseWithoutResolving(s: String) = {
