+++ sconfig/jvm/src/test/scala/org/ekrich/config/impl/EquivalentsTest.scala
@@ -9,6 +9,7 @@ import org.junit.Assert._
 import org.junit._
 
 import org.ekrich.config._
+import FileUtils._
 
 class EquivalentsTest extends TestUtils {
   private def equivDirs() = {
