+++ sconfig/shared/src/test/scala/org/ekrich/config/impl/TokenTest.scala
